+++ webpack.config.js
@@ -6,6 +6,14 @@ module.exports = {
     libraryTarget: 'umd',
   },
 
+  module: {
+    loaders: [{
+      test: /\.jsx?$/,
+      exclude: /node_modules/,
+      loader: 'babel-loader',
+    }],
+  },
+
   externals: [
     {
       'react': {